issues/30: move indicator components to component directory
