+++ src/component/arrivalIndicator/Indicator.css
@@ -1,0 +1,7 @@
+.arrival-estimated-late {
+    color: red;
+}
+
+.arrival-estimated-early {
+    color: yellowgreen;
+}
